Refactor ComingEpisodes.py -> coming_episodes.py
